+++ docs/participant_transfer/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to this gear are documented in this file.
 
+## 1.1.1
+* Adds support for affiliated studies that doesn't have a separate group in Flywheel
+
 ## 1.1.0
 * Updates to Python 3.12 and switches to use `fw-gear` instead of `flywheel-gear-toolkit` (now deprecated)